implementation to secure unique ADMIN data

+++ src/main/java/pl/codeschool/controller/AdminDeleteUserController.java
@@ -1,6 +1,10 @@
 package pl.codeschool.controller;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import pl.codeschool.dao.UserDao;
+import pl.codeschool.model.Admin;
+import pl.codeschool.model.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,6 +16,8 @@ import java.io.IOException;
 @WebServlet("/admin/delete/user")
 public class AdminDeleteUserController extends HttpServlet {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(AdminDeleteUserController.class);
+
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html; charset=UTF-8");
 
@@ -20,10 +26,14 @@ public class AdminDeleteUserController extends HttpServlet {
         if (paramUserId != null && !"".equals(paramUserId)) {
             try {
                 int userId = Integer.parseInt(paramUserId);
-                UserDao.delete(userId);
+                User foundedUser = UserDao.read(userId);
+                if (foundedUser != null && !foundedUser.getUserName().equals(Admin.getAdminUsername())) {
+                    UserDao.delete(userId);
+                } else
+                    request.setAttribute("tryingDeleteSuperAdmin", true);
 
             } catch (NumberFormatException e) {
-                e.printStackTrace();
+                LOGGER.info(e.getMessage());
             }
         }
         request.getRequestDispatcher("/admin/users")
